--- ftp_project/__init__.py
@@ -1,1 +1,0 @@
-# FTP Project - Raw Socket Implementation
