--- ftp_project/clamav_agent/__init__.py
@@ -1,1 +1,0 @@
-# ClamAV Agent Package
