--- ftp_project/client/__init__.py
@@ -1,1 +1,0 @@
-# FTP Client Package
