--- ftp_project/client/core/__init__.py
@@ -1,1 +1,0 @@
-# Core FTP functionality
