+++ ftp_project/client/core/virus_scan.py
@@ -1,8 +1,6 @@
-import os
-import subprocess
 import socket
-import json
-import time
+import struct
+import os
 from .config import Config
 from .utils import Utils
 import logging
