--- ftp_project/client/networking/__init__.py
@@ -1,1 +1,0 @@
-# Networking components
