--- ftp_project/client/ui/__init__.py
@@ -1,1 +1,0 @@
-# UI components
