+++ ftp_project/client/ui/login_window.py
@@ -1,9 +1,9 @@
 import tkinter as tk
-from tkinter import messagebox, ttk
+from tkinter import ttk, messagebox
 import threading
-import time
-import os
 import sys
+import os
+import socket
 
 # Thêm thư mục hiện tại vào Python path
 sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
