--- ftp_project/demo.py
@@ -1,114 +1,0 @@
-#!/usr/bin/env python3
-"""
-Demo Raw Socket FTP Implementation - Restructured Project
-"""
-
-import sys
-import os
-
-# Add current directory to Python path
-sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
-
-from client.core.raw_socket_ftp import *
-from client.core.ftp_command import FTPCommands
-
-def demo_new_structure():
-    """Demo the new project structure"""
-    print("=== FTP PROJECT - NEW STRUCTURE DEMO ===\n")
-    
-    print("📁 Project Structure:")
-    print("""
-ftp_project/
-├── client/
-│   ├── core/          # Core FTP logic
-│   ├── ui/            # GUI components  
-│   ├── networking/    # Network utilities
-│   └── downloads/     # Download directory
-├── clamav_agent/      # Virus scanning
-├── tests/             # Test files
-└── README.md          # Documentation
-    """)
-    
-    print("🔧 Core Components:")
-    print("  ✅ raw_socket_ftp.py - Raw socket implementation")
-    print("  ✅ ftp_command.py - Command line interface")
-    print("  ✅ ftp_helpers.py - Helper functions")
-    print("  ✅ virus_scan.py - Virus scanning")
-    print("  ✅ config.py - Configuration")
-    print("  ✅ utils.py - Utilities")
-    
-    print("\n🎯 UI Components:")
-    print("  ✅ ftp_gui.py - Main GUI")
-    print("  ✅ login_window.py - Login dialog")
-    print("  ✅ main.py - GUI entry point")
-    
-    print("\n🌐 Networking:")
-    print("  ✅ client.py - Network client")
-    
-    print("\n🦠 ClamAV Agent:")
-    print("  ✅ handler.py - Request handler")
-    print("  ✅ scanner.py - Virus scanner")
-    print("  ✅ sever_clam.py - ClamAV server")
-    print("  ✅ main.py - Agent entry point")
-    
-    # Test import
-    try:
-        ftp = FTP()
-        client = FTPCommands(ftp)
-        print(f"\n✅ Successfully initialized FTP client!")
-        print(f"📋 Available commands: {len([cmd for cmd in dir(client) if cmd.startswith('do_')])}")
-        
-    except Exception as e:
-        print(f"\n❌ Error initializing client: {e}")
-    
-    print("\n🚀 How to run:")
-    print("  python run_client.py          # CLI mode (default)")
-    print("  python run_client.py --cli    # CLI mode")
-    print("  python run_client.py --gui    # GUI mode")
-    print("  python demo.py                # This demo")
-    
-    print("\n✨ Features:")
-    print("  • Raw socket FTP implementation")
-    print("  • No ftplib dependency")
-    print("  • Modular structure")
-    print("  • Easy to maintain")
-    print("  • Clean separation of concerns")
-
-def test_imports():
-    """Test all imports work correctly"""
-    print("\n=== IMPORT TESTS ===")
-    
-    try:
-        from client.core.raw_socket_ftp import FTP
-        print("✅ Core FTP import: OK")
-    except ImportError as e:
-        print(f"❌ Core FTP import: {e}")
-    
-    try:
-        from client.core.ftp_command import FTPCommands
-        print("✅ FTP Commands import: OK")
-    except ImportError as e:
-        print(f"❌ FTP Commands import: {e}")
-    
-    try:
-        from client.core.utils import Utils
-        print("✅ Utils import: OK")
-    except ImportError as e:
-        print(f"❌ Utils import: {e}")
-    
-    try:
-        from client.core.config import Config
-        print("✅ Config import: OK")
-    except ImportError as e:
-        print(f"❌ Config import: {e}")
-    
-    print("\n🎉 All core imports successful!")
-
-if __name__ == "__main__":
-    demo_new_structure()
-    test_imports()
-    
-    print("\n" + "="*50)
-    print("🎉 PROJECT RESTRUCTURE COMPLETE!")
-    print("🚀 Ready to use with new organized structure!")
-    print("="*50)
