--- tests/pytest.ini
@@ -1,26 +1,0 @@
-[pytest]
-# Test discovery
-testpaths = .
-python_files = test_*.py
-python_classes = Test*
-python_functions = test_*
-
-# Test markers
-markers =
-    session: Session management tests
-    file_ops: File operation tests
-    directory_ops: Directory operation tests
-    local_ops: Local operation tests
-    virus_scan: Virus scanning tests
-    integration: Integration tests (require servers)
-    real_server: Real server integration tests (require actual FTP server and credentials)
-
-# Output options
-addopts = 
-    -v
-    --tb=short
-    --strict-markers
-    --strict-config
-
-# Minimum pytest version
-minversion = 7.0
