--- tests/requirements.txt
@@ -1,2 +1,0 @@
-# Testing framework
-pytest>=7.0.0
